+++ Sources/Sentry/SentryHttpTransport.m
@@ -17,7 +17,10 @@
 #import "SentryNSURLRequest.h"
 #import "SentryNSURLRequestBuilder.h"
 #import "SentryOptions.h"
+#import "SentryReachability.h"
 #import "SentrySerialization.h"
+
+static NSTimeInterval const cachedEnvelopeSendDelay = 0.1;
 
 @interface
 SentryHttpTransport ()
@@ -30,6 +33,7 @@ SentryHttpTransport ()
 @property (nonatomic, strong) SentryEnvelopeRateLimit *envelopeRateLimit;
 @property (nonatomic, strong) SentryDispatchQueueWrapper *dispatchQueue;
 @property (nonatomic, strong) dispatch_group_t dispatchGroup;
+@property (nonatomic, strong) SentryReachability *reachability;
 
 /**
  * Relay expects the discarded events split by data category and reason; see
@@ -59,6 +63,7 @@ SentryHttpTransport ()
               rateLimits:(id<SentryRateLimits>)rateLimits
        envelopeRateLimit:(SentryEnvelopeRateLimit *)envelopeRateLimit
     dispatchQueueWrapper:(SentryDispatchQueueWrapper *)dispatchQueueWrapper
+            reachability:(SentryReachability *)reachability
 {
     if (self = [super init]) {
         self.options = options;
@@ -74,10 +79,30 @@ SentryHttpTransport ()
         self.discardedEvents = [NSMutableDictionary new];
         [self.envelopeRateLimit setDelegate:self];
         [self.fileManager setDelegate:self];
+        self.reachability = reachability;
 
         [self sendAllCachedEnvelopes];
+
+#if !TARGET_OS_WATCH
+        [self.reachability monitorURL:[NSURL URLWithString:@"https://sentry.io"]
+                        usingCallback:^(BOOL connected, NSString *_Nonnull typeDescription) {
+                            if (connected) {
+                                SENTRY_LOG_DEBUG(@"Internet connection is back.");
+                                [self sendAllCachedEnvelopes];
+                            } else {
+                                SENTRY_LOG_DEBUG(@"Lost internet connection.");
+                            }
+                        }];
+#endif
     }
     return self;
+}
+
+- (void)dealloc
+{
+#if !TARGET_OS_WATCH
+    [self.reachability stopMonitoring];
+#endif
 }
 
 - (void)sendEnvelope:(SentryEnvelope *)envelope
@@ -126,22 +151,31 @@ SentryHttpTransport ()
 
 - (BOOL)flush:(NSTimeInterval)timeout
 {
+    // Calculate the dispatch time of the flush duration as early as possible to guarantee an exact
+    // flush duration. Any code up to the dispatch_group_wait can take a couple of ms, adding up to
+    // the flush duration.
+    dispatch_time_t delta = (int64_t)(timeout * (NSTimeInterval)NSEC_PER_SEC);
+    dispatch_time_t dispatchTimeout = dispatch_time(DISPATCH_TIME_NOW, delta);
+
+    // Double-Checked Locking to avoid acquiring unnecessary locks.
+    if (_isFlushing) {
+        SENTRY_LOG_DEBUG(@"Already flushing.");
+        return NO;
+    }
+
     @synchronized(self) {
         if (_isFlushing) {
-            SENTRY_LOG_DEBUG(@"SentryHttpTransport: Already flushing.");
+            SENTRY_LOG_DEBUG(@"Already flushing.");
             return NO;
         }
 
-        SENTRY_LOG_DEBUG(@"SentryHttpTransport: Start flushing.");
+        SENTRY_LOG_DEBUG(@"Start flushing.");
 
         _isFlushing = YES;
         dispatch_group_enter(self.dispatchGroup);
     }
 
     [self sendAllCachedEnvelopes];
-
-    dispatch_time_t delta = (int64_t)(timeout * (NSTimeInterval)NSEC_PER_SEC);
-    dispatch_time_t dispatchTimeout = dispatch_time(DISPATCH_TIME_NOW, delta);
 
     intptr_t result = dispatch_group_wait(self.dispatchGroup, dispatchTimeout);
 
@@ -150,10 +184,10 @@ SentryHttpTransport ()
     }
 
     if (result == 0) {
-        SENTRY_LOG_DEBUG(@"SentryHttpTransport: Finished flushing.");
+        SENTRY_LOG_DEBUG(@"Finished flushing.");
         return YES;
     } else {
-        SENTRY_LOG_DEBUG(@"SentryHttpTransport: Flushing timed out.");
+        SENTRY_LOG_DEBUG(@"Flushing timed out.");
         return NO;
     }
 }
@@ -207,10 +241,11 @@ SentryHttpTransport ()
 
 - (void)sendAllCachedEnvelopes
 {
+    SENTRY_LOG_DEBUG(@"sendAllCachedEnvelopes start.");
+
     @synchronized(self) {
         if (self.isSending || ![self.requestManager isReady]) {
-            [SentryLog logWithMessage:@"SentryHttpTransport: Already sending."
-                             andLevel:kSentryLevelDebug];
+            SENTRY_LOG_DEBUG(@"Already sending.");
             return;
         }
         self.isSending = YES;
@@ -218,8 +253,7 @@ SentryHttpTransport ()
 
     SentryFileContents *envelopeFileContents = [self.fileManager getOldestEnvelope];
     if (nil == envelopeFileContents) {
-        [SentryLog logWithMessage:@"SentryHttpTransport: No envelopes left to send."
-                         andLevel:kSentryLevelDebug];
+        SENTRY_LOG_DEBUG(@"No envelopes left to send.");
         [self finishedSending];
         return;
     }
@@ -254,11 +288,11 @@ SentryHttpTransport ()
 
 - (void)deleteEnvelopeAndSendNext:(NSString *)envelopePath
 {
-    [SentryLog logWithMessage:@"SentryHttpTransport: Deleting envelope and sending next."
-                     andLevel:kSentryLevelDebug];
+    SENTRY_LOG_DEBUG(@"Deleting envelope and sending next.");
     [self.fileManager removeFileAtPath:envelopePath];
     self.isSending = NO;
-    [self sendAllCachedEnvelopes];
+    [self.dispatchQueue dispatchAfter:cachedEnvelopeSendDelay
+                                block:^{ [self sendAllCachedEnvelopes]; }];
 }
 
 - (void)sendEnvelope:(SentryEnvelope *)envelope
@@ -278,8 +312,7 @@ SentryHttpTransport ()
                 [_self.rateLimits update:response];
                 [_self deleteEnvelopeAndSendNext:envelopePath];
             } else {
-                [SentryLog logWithMessage:@"SentryHttpTransport: No internet connection."
-                                 andLevel:kSentryLevelDebug];
+                SENTRY_LOG_DEBUG(@"No internet connection.");
                 [_self finishedSending];
             }
         }];
@@ -287,11 +320,11 @@ SentryHttpTransport ()
 
 - (void)finishedSending
 {
-    SENTRY_LOG_DEBUG(@"SentryHttpTransport: Finished sending.");
+    SENTRY_LOG_DEBUG(@"Finished sending.");
     @synchronized(self) {
         self.isSending = NO;
         if (self.isFlushing) {
-            SENTRY_LOG_DEBUG(@"SentryHttpTransport: Stop flushing.");
+            SENTRY_LOG_DEBUG(@"Stop flushing.");
             self.isFlushing = NO;
             dispatch_group_leave(self.dispatchGroup);
         }