+++ Tests/SentryTests/TestUtils/SentryTestObserver.m
@@ -27,24 +27,20 @@ SentryTestObserver ()
 
 @implementation SentryTestObserver
 
-#if defined(TESTCI)
 + (void)load
 {
+#if defined(TESTCI)
     [[XCTestObservationCenter sharedTestObservationCenter]
         addTestObserver:[[SentryTestObserver alloc] init]];
-}
-#elif defined(TEST)
-+ (void)load
-{
+#endif
     [SentryLog configure:YES diagnosticLevel:kSentryLevelDebug];
 }
-#endif
 
 - (instancetype)init
 {
     if (self = [super init]) {
         SentryOptions *options = [[SentryOptions alloc] init];
-        options.dsn = @"https://a92d50327ac74b8b9aa4ea80eccfb267@o447951.ingest.sentry.io/5428557";
+        options.dsn = @"https://6cc9bae94def43cab8444a99e0031c28@o447951.ingest.sentry.io/5428557";
         options.environment = @"unit-tests";
         options.debug = YES;
         options.enableAutoSessionTracking = NO;
@@ -52,21 +48,18 @@ SentryTestObserver ()
 
         // The SentryCrashIntegration enriches the scope. We need to install the integration
         // once to get the scope data.
+        [SentrySDK startWithOptionsObject:options];
 
-        // When running the SentryTestObserver the code gets stuck when accessing the
-        // UIScreen.mainScreen in SentryCrashIntegration. We disable adding extra context for now.
-        // Ideally we somehow check here if we can access UIScreen.mainScreen, see
-        // https://github.com/getsentry/sentry-cocoa/issues/2284
-        //        [SentrySDK startWithOptionsObject:options];
-        //
-        //        self.scope = [[SentryScope alloc] init];
-        //        [SentryCrashIntegration enrichScope:self.scope
-        //                               crashWrapper:[SentryCrashWrapper sharedInstance]];
+        self.scope = [[SentryScope alloc] init];
+        [SentryCrashIntegration enrichScope:self.scope
+                               crashWrapper:[SentryCrashWrapper sharedInstance]];
 
         self.options = options;
     }
     return self;
 }
+
+#pragma mark - XCTestObservation
 
 - (void)testCaseWillStart:(XCTestCase *)testCase
 {
@@ -76,6 +69,11 @@ SentryTestObserver ()
     // The tests might have a different time set
     [crumb setTimestamp:[NSDate new]];
     [self.scope addBreadcrumb:crumb];
+}
+
+- (void)testBundleDidFinish:(NSBundle *)testBundle
+{
+    [SentrySDK flush:5.0];
 }
 
 - (void)testCase:(XCTestCase *)testCase didRecordIssue:(XCTIssue *)issue
@@ -96,12 +94,6 @@ SentryTestObserver ()
     [hub captureException:exception withScope:hub.scope];
 
     [SentryCurrentDate setCurrentDateProvider:currentDateProvider];
-}
-
-- (void)testBundleDidFinish:(NSBundle *)testBundle
-{
-    // Wait for events to flush out.
-    [NSThread sleepForTimeInterval:3.0];
 }
 
 @end